requirements.txt
appdirs==1.4.3
facebook-sdk==2.0.0
packaging==16.8
pkg-resources==0.0.0
pluggy==0.4.0
py==1.4.33
pyparsing==2.2.0
pytest==3.1.0
requests==2.14.2
six==1.10.0
tinydb==3.2.3
virtualenv==15.1.0
